22-holding-weapon
I also made it so when picking up the weapon it will jump the gameobject in left hand put this will work differently in the switching weapons part. Also F Key now drops gameobjects.

Also two new gameobjects(cubes) are added for the weapons to test around. Gameobjecs also have boolean to check if the are left or right hand objects.